Create endpoints for websockets connections and its router.
