+++ main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from db.database import engine
 from db import models
-from routers import user_router, product_router, bid_router, user_address_router
+from routers import user_router, product_router, bid_router, user_address_router, websocket_router
 from auth import authentication
 
 
@@ -13,6 +13,7 @@ app.include_router(product_router.router)
 app.include_router(authentication.router)
 app.include_router(user_router.router)
 app.include_router(user_address_router.router)
+app.include_router(websocket_router.router)
 
 
 @app.get('/')
