+++ notifications/notification.py
@@ -32,6 +32,7 @@ class InAppNotification:
 
     async def connect(self, websocket: WebSocket):
         await websocket.accept()
+        print("new websocket added")
         self.active_connections.append(websocket)
 
     def disconnect(self, websocket: WebSocket):
